Reformat WalletModelTest
